refactor: delete buttons and text index.js
Also simplify project structure
