--- src/components/common/buttons/index.js
@@ -1,7 +1,0 @@
-import PrimaryButton from './primary-button.js';
-import SecondaryButton from './secondary-button.js';
-
-export {
-    PrimaryButton,
-    SecondaryButton
-};
