+++ src/components/common/h1-text.js
