+++ src/components/common/h2-text.js
