+++ src/components/common/index.js
@@ -1,13 +1,12 @@
-import {
-    H1Text,
-    H2Text,
-    ModalH1Text,
-    SubHeaderText,
-} from './texts/index.js';
 
-import { PrimaryButton, SecondaryButton } from './buttons';
 import AppSpinner from './app-spinner';
+import H1Text from './h1-text';
+import H2Text from './h2-text';
 import LanguageSelector from './language-selector';
+import ModalH1Text from './modal-h1-text';
+import PrimaryButton from './primary-button';
+import SecondaryButton from './secondary-button';
+import SubHeaderText from './sub-header-text';
 
 export {
     AppSpinner,
