--- src/components/common/texts/index.js
@@ -1,11 +1,0 @@
-import H1Text from './h1-text.js';
-import H2Text from './h2-text.js';
-import ModalH1Text from './modal-h1-text.js';
-import SubHeaderText from './sub-header-text.js';
-
-export {
-    H1Text,
-    H2Text,
-    SubHeaderText,
-    ModalH1Text,
-};
